alibaba/webhook-handler/requirements.txt
# Alibaba Cloud SDK
tablestore>=6.3.0
aliyun-mns>=1.1.5
dashscope>=1.20.0

# Telegram
python-telegram-bot>=20.0

# Audio processing
openai>=1.0.0
pydub>=0.25.0

# Utilities
pytz>=2024.1
httpx>=0.25.0
